requirements.txt
streamlit==1.42.0
anthropic==0.49.0
scikit-learn==1.6.1
plotly==5.24.1
pandas==2.2.3
numpy==1.26.4
fpdf2==2.8.1
python-dotenv==1.0.1
